JBPM 4714: relocated UpdateTimerCommand and divided UpdateTaskNameCommand into SetNodeNameCommand and SetTaskPropertyCommand
