+++ jbpm-test-coverage/pom.xml
@@ -63,6 +63,10 @@
     </dependency>
     <dependency>
       <groupId>org.jbpm</groupId>
+      <artifactId>jbpm-human-task-core</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.jbpm</groupId>
       <artifactId>jbpm-case-mgmt</artifactId>
     </dependency>
     <dependency>
